test(runtime): isolate managed docs contract
